+++ crates/nextchat-communication/src/incoming/friend_request.rs
@@ -1,5 +1,6 @@
 //! NextChat Communication friend request event module.
 
+use async_trait::async_trait;
 use nextchat_database::Client;
 
 use crate::{CommunicationMessage, Connection, StorageType};
@@ -8,8 +9,9 @@ use super::PacketEvent;
 
 pub struct FriendRequestEvent;
 
+#[async_trait]
 impl PacketEvent for FriendRequestEvent {
-    fn run(
+    async fn run(
         _connection: &Connection,
         message: &CommunicationMessage,
         _client: &Client,