+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased (0.1.0-alpha1)
 
+### 23/03/2021
+-   Add `async-trait` for incoming events (NextChat Communication).
+
 ### 22/03/2021
 -   Add NextChat Communication module.
 -   Add a new modular structure.